Major changes

- Removed String concatenation based result storage, replaced with Jena graph
- Changes in benchmark result: fixed typos, removed term collisions, introduced new edges
- Replaced positional arguments based worker configuration to key value pair based arguments
- Revamped result processing for SPARQLWorker(Processing in background threads, caching of already processed results)
- Added feature to override response type of SPARQL query from the configuration
- Improved precision of time measurement from microseconds to nanoseconds
- New class hierarchy for Worker classes
- Several bug fixes

+++ iguana.resultprocessor/src/main/java/org/aksw/iguana/rp/metrics/impl/EachQueryMetric.java
@@ -56,7 +56,7 @@ public class EachQueryMetric extends AbstractMetric {
 		
 		LOGGER.debug(this.getShortName() + " has received " + p);
 
-		long time = (long) p.get(COMMON.RECEIVE_DATA_TIME);
+		double time = (double) p.get(COMMON.RECEIVE_DATA_TIME);
 		Boolean success = (Boolean) (((long) p.get(COMMON.RECEIVE_DATA_SUCCESS))>0?true:false);
 		String queryID = p.getProperty(COMMON.QUERY_ID_KEY);
 		long err = (long) p.get(COMMON.RECEIVE_DATA_SUCCESS);
